requirements.txt
selenium==4.1.3
webdriver-manager==3.5.4
pytest==7.1.2
pyyaml==6.0
pytest-html==3.1.1
pytest-dotenv==1.0.0
